add PMPML_32, PMPML_64
yet unseeded.
unconst'ed random_data_for_MPSHF to prepare setting the seed.

also added SMHasher --listnames option.

+++ PMP_Multilinear_test.h
@@ -58,10 +58,10 @@ void PMPML_32_CPP( const void * key, int len, uint32_t seed, void * res );
 void PMPML_32_CPP_randomize();
 
 #if (defined _WIN64) || (defined __x86_64__)
-void PMPML_64_CPP( const void * key, int len, uint64_t seed, void * res );
+void PMPML_64_CPP( const void * key, int len, uint32_t seed, void * res );
 void PMPML_64_CPP_randomize();
 
-void PMPML_64_CPP_out_32( const void * key, int len, uint64_t seed, void * res );
+void PMPML_64_CPP_out_32( const void * key, int len, uint32_t seed, void * res );
 void PMPML_64_CPP_out_32_randomize();
 int PMPML_TestSpeedAlt_64_out_32( const void * key, int len, int iter );
 #endif
